requirements.txt
# chapter 1
openai>=2.14.0
anthropic>=0.75.0
google-generativeai>=0.8.0
tiktoken==0.7.0
streamlit>=1.33.0
pydantic>=2.12.5
langchain>=1.2.0
langchain-community>=0.4.1
langchain-openai>=1.1.6
langchain-core>=1.2.5
langchain-google-genai>=4.1.2
langchain-anthropic>=1.3.0
langchain-text-splitters>=1.1.0
langgraph>=1.0.5
langgraph-prebuilt>=1.0.5
langsmith>=0.5.1
python-dotenv==1.0.0

## chapter 5
# Webサイト要約アプリ用
requests>=2.32.5
beautifulsoup4==4.12.3

# Youtube要約アプリ用
youtube-transcript-api==1.2.3
pytube==15.0.0

# chapter 7
PyMuPDF==1.23.25
faiss-cpu==1.7.4

# chapter 8
langsmith>=0.3.45

# chapter 9
duckduckgo_search==4.2
html2text==2020.1.16
lxml[html_clean]==5.2.1
readability-lxml==0.8.1

# chapter 10
streamlit-feedback==0.1.3

# chapter 11
python-magic==0.4.27
google-cloud-bigquery==3.21.0
db-dtypes==1.2.0
